backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
google-generativeai==0.3.0
langchain==0.1.0
langgraph==0.0.1
pinecone-client==3.0.0
openai==1.3.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.0
textblob==0.17.1
langdetect==1.0.9
redis==5.0.1
python-multipart==0.0.6
aiohttp==3.9.1